move core component types to core folder
