+++ components/core/Card/__tests__/card.spec.tsx
@@ -6,7 +6,8 @@ import { Card } from '../Card';
 import { CARD_PROPS_MOCKS } from '../__mocks__/card.mocks';
 
 import { renderWithTheme } from '../../../../helpers/Theme';
-import { CardType } from '../../../../helpers/types';
+
+import { CardType } from '../../types';
 
 it('should render standart card', () => {
   const { container } = render(
