+++ components/core/CustomLink/CustomLink.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode } from 'react';
 
 import { StyledCustomLink } from './styles';
 
-import { CustomLinkSize } from '../../../helpers/types';
+import { CustomLinkSize } from '../types';
 
 type CustomLinkProps = {
   className?: string;
