+++ components/core/Logo/__stories__/logo.story.tsx
@@ -5,7 +5,7 @@ import { select } from '@storybook/addon-knobs';
 
 import { Logo } from '../Logo';
 
-import { LogoSize } from '../../../../helpers/types';
+import { LogoSize } from '../../types';
 
 const stories = storiesOf('Logo', module);
