+++ components/ui/FullLoader/FullLoader.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode } from 'react';
 
 import { StyledFullLoaderWrapper } from './styles';
 
-import { LoaderType } from '../../../helpers/types';
+import { LoaderType } from '../../core/types';
 
 type FullLoaderProps = {
   isShowed?: boolean;
