+++ helpers/types.tsx
@@ -2,39 +2,6 @@ import { ReactNode } from 'react';
 
 // Core / ui components
 
-export enum LogoSize {
-  BIG = 'big',
-  SMALL = 'small',
-}
-
-export enum LoaderType {
-  CIRCLE = 'circle',
-  LINE = 'line',
-}
-
-export enum LoaderMode {
-  DARK = 'dark',
-  LIGHT = 'light',
-}
-
-export type ListItem = {
-  iconSrc?: string;
-  id: number;
-  name: string;
-  path: string;
-};
-
-export enum CardType {
-  SUGGESTION = 'suggestion',
-  STANDART = 'standart',
-  HIGHLIGHT = 'highlight',
-}
-
-export enum CustomLinkSize {
-  BIG = 'big',
-  SMALL = 'small',
-}
-
 // Restaurant data
 
 export type Location = {
